requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0

# Image Processing
opencv-python-headless>=4.5.0
Pillow>=8.3.0

# Machine Learning
tensorflow>=2.6.0,<3.0.0
scikit-learn>=0.24.0
scipy>=1.7.0

# Development Tools
pytest>=6.2.0
black>=21.7b0
flake8>=3.9.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Testing & Evaluation
pytest-cov>=2.12.0

# Utilities
python-dotenv>=0.19.0
